add ForgetPage
