+++ src/components/reviewPage/ReviewCardList.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useInfiniteQuery } from "@tanstack/react-query";
-import { useInView } from "react-intersection-observer"; // 用於觀察元素是否進入視窗
+import { useInView } from "react-intersection-observer";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSpinner } from "@fortawesome/free-solid-svg-icons";
 import ReviewCard from "./ReviewCard";
